+++ Project1/DirectX12/Util/CharToWChar.cpp
@@ -7,7 +7,7 @@
 size_t ToWChar(wchar_t** ppBuf, size_t charWordLen, const char* pCharWord, size_t ppBufMaxCount)
 {
 	size_t rtn = 0;
-	int size = charWordLen;
+	unsigned int size = static_cast<unsigned int>(charWordLen);
 	*ppBuf = new wchar_t[size + 10];
 	mbstowcs_s(&rtn, *ppBuf, charWordLen, pCharWord, _TRUNCATE);
 
@@ -27,7 +27,7 @@ size_t ToWChar(wchar_t** ppBuf, std::string cstr)
 size_t ToChar(char** ppBuf, size_t ppBufSizeInByte, const wchar_t* wcStr, size_t convertWordNum)
 {
 	size_t rtn = 0;
-	int size = convertWordNum;
+	unsigned int size = static_cast<unsigned int>(convertWordNum);
 	*ppBuf = new char[size];
 	wcstombs_s(&rtn, *ppBuf, ppBufSizeInByte, wcStr, _TRUNCATE);
